Moved everything into the Framework namespace and provided an AdapterInterface to enhance readability for developers
